refactor: rename coding-tools → system-tools
- coding-tools.ts → system-tools.ts (root + packages/tools)
- CodingToolName → SystemToolName
- createCodingTools → createSystemTools
- Backward-compatible alias: createCodingTools still exported (re-export)
- Updated all imports: engine.ts, completions.ts, index.ts, tests

1153/1153 tests passing.
